fix problems
